refactor treeshakable functionality in npm module
