+++ auxpack-npm/utils/parser.js
@@ -1,4 +1,35 @@
 module.exports = (stats) => {
+  const moduleTypes = {
+    'harmony side effect evaluation': 'esm',
+    'harmony import specifier': 'esm',
+    'cjs require': 'cjs'
+  }
+
+  const obj = {
+    cjs: [],
+    esm: [],
+    both: []
+  }
+
+  stats.modules.forEach(module => {
+    const { name, size, reasons} = module;
+    const isEsm = reasons.some(reason => moduleTypes[reason.type] === 'esm');
+    const isCjs = reasons.some(reason => moduleTypes[reason.type] === 'cjs');
+    
+    const reasonTypes = reasons.map(reason => {
+      return { name: reason.module, type: reason.type}
+    })
+
+    const moduleWithTypes = {
+      name, 
+      size,
+      reasonTypes
+    }
+
+    if (obj['esm'] && isEsm && !isCjs) obj['esm'].push(moduleWithTypes);
+    else if (obj['both'] && isEsm && isCjs) obj['both'].push(moduleWithTypes);
+    else if (obj['cjs'] && !isEsm && isCjs) obj['cjs'].push(moduleWithTypes);
+  })
 
   return {
     timeStamp: Date.now(),
@@ -24,6 +55,9 @@ module.exports = (stats) => {
             id: module.id
           }))
         : []
-    }))
+    })),
+
+    treeStats: obj
   };
 };
+
